+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -3080,6 +3080,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("MissingExternalMemberInjector.kt")
+      public void testMissingExternalMemberInjector() {
+        run("MissingExternalMemberInjector.kt");
+      }
+
+      @Test
       @TestMetadata("MultibindingEmptyDaggerSetProvider.kt")
       public void testMultibindingEmptyDaggerSetProvider() {
         run("MultibindingEmptyDaggerSetProvider.kt");